src/python/requirements.txt
PyGithub==1.44
jinja2==2.10.3